Adding confirm message to users delete view

+++ ajuda.me/resources/views/users/index.blade.php
@@ -29,7 +29,7 @@
                                         <a href="/users/{{$user->id}}/edit" style="background-color:#00FF7F;"class="btn btn-warning">Update</a>
                                     </td>
                                     <td>
-                                        <a href="/users/{{$user->id}}/delete" style="background-color:#FF6347;"class="btn btn-danger">Delete</a>
+                                        <a href="/users/{{$user->id}}/delete" onclick="return confirm('Voce tem certeza ?')" style="background-color:#FF6347;"class="btn btn-danger">Delete</a>
                                     </td>
                                 </tr>
                             @endforeach
